+++ .gitignore
@@ -5,3 +5,4 @@ node_modules/
 dist/
 build/
 .firebase/
+src/javascripts/helpers/apiKeys.json
