+++ db/base.json
@@ -1,0 +1,3 @@
+{
+  "projects": "projects"
+}
