+++ src/index.html
@@ -7,7 +7,7 @@
     <title>John Thielman</title>
     <!-- <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"> -->
     <link href="https://fonts.googleapis.com/css?family=Jim+Nightshade|Lancelot|Macondo+Swash+Caps|Milonga|Uncial+Antiqua|UnifrakturCook:700|UnifrakturMaguntia&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="./main.css">
+    <!-- <link rel="stylesheet" href="./styles/main.scss"> -->
 </head>
 <body>
     <nav>
@@ -19,18 +19,20 @@
         </ul>
     </nav>
 
-    <div class="fullPage" id="bioPage">
-        <p>I have degrees in Classics and Historical Linguistics, and an abiding interest in mathematics and similar styles of problem solving.  Development is great because it combines my interests in syntax, logical systems and problem solving, and then it throws in a dash of creativity.  In short, it's both challenging and fun.</p>
-    </div>
-    <div class="fullPage hide" id="technologiesPage">
-        <ul>
-            <li>HTML</li>
-            <li>CSS</li>
-            <li>JavaScript</li>
-        </ul>
-    </div>
-    <div class="fullPage hide" id="projectsPage">
-        <div class="container" id="projectsContainer"></div>
+    <div class="container">
+        <div class="fullPage" id="bioPage">
+            <p>I have degrees in Classics and Historical Linguistics, and an abiding interest in mathematics and similar styles of problem solving.  Development is great because it combines my interests in syntax, logical systems and problem solving, and then it throws in a dash of creativity.  In short, it's both challenging and fun.</p>
+        </div>
+        <div class="fullPage hide" id="technologiesPage">
+            <ul>
+                <li>HTML</li>
+                <li>CSS</li>
+                <li>JavaScript</li>
+            </ul>
+        </div>
+        <div class="fullPage hide" id="projectsPage">
+            <div class="container" id="projectsContainer"></div>
+        </div>
     </div>
 
     <footer class="page-footer">
@@ -40,6 +42,6 @@
         <a href="https://twitter.com/dschonn">Twitter</a>
     </footer>
 
-    <script src="./projects.js"></script>
+    <!-- <script src="./javascripts/main.js"></script> -->
 </body>
 </html>
