+++ src/javascripts/helpers/apiKeys.example.json
@@ -1,0 +1,12 @@
+{
+  "firebaseKeys": {
+    "apiKey": "",
+    "authDomain": "",
+    "databaseURL": "",
+    "projectId": "",
+    "storageBucket": "",
+    "messagingSenderId": "",
+    "appId": "",
+    "measurementId": ""
+  }
+}
